pyproject.toml
[project]
name = "kiwi-tcms-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "muuuuwa", email = "muuuuwa@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tcms-api>=13.3",
]
classifiers = ["Private :: Do Not Upload"]

[project.scripts]
runserver = "kiwi_tcms_mcp_server.kiwi_tcms_server:runserver"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
